requirements-dev.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.0.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
peewee==3.15.3
Pillow==9.3.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.6
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.2.0
